Extracted calculations/add.
