+++ src/duration/calculations.js
@@ -1,5 +1,16 @@
 import {
-  UNIT_NAMES,
+  addMicroseconds,
+  addMilliseconds,
+  addSeconds,
+  addMinutes,
+  addHours,
+  addDays,
+  addWeeks,
+  addMonths,
+  addYears,
+} from './calculations/add';
+
+import {
   ONE_HOUR,
   ONE_MINUTE,
   ONE_SECOND,
@@ -36,33 +47,6 @@ const addUnit = (isoString, amount, unit) => {
   return toIso(fragments);
 };
 
-export const addMilliseconds = curry((amount, isoString) =>
-  addUnit(isoString, amount / (10 ** 3), UNIT_NAMES.seconds));
-
-export const addMicroseconds = curry((amount, isoString) =>
-  addUnit(isoString, amount / (10 ** 6), UNIT_NAMES.seconds));
-
-export const addSeconds = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.seconds));
-
-export const addMinutes = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.minutes));
-
-export const addHours = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.hours));
-
-export const addDays = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.days));
-
-export const addWeeks = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.weeks));
-
-export const addMonths = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.months));
-
-export const addYears = curry((amount, isoString) =>
-  addUnit(isoString, amount, UNIT_NAMES.years));
-
 export const subtractMilliseconds = curry((amount, isoString) =>
   addMilliseconds(amount * -1, isoString));
 
@@ -89,3 +73,15 @@ export const subtractMonths = curry((amount, isoString) =>
 
 export const subtractYears = curry((amount, isoString) =>
   addYears(amount * -1, isoString));
+
+export {
+  addMicroseconds,
+  addMilliseconds,
+  addSeconds,
+  addMinutes,
+  addHours,
+  addDays,
+  addWeeks,
+  addMonths,
+  addYears,
+} from './calculations/add';
